Decouple VPN helpers
Store and notifications.
