+++ app-android/app/src/main/kotlin/com/algoritmico/passepartout/PassepartoutVpnService.kt
@@ -4,41 +4,22 @@
 
 package com.algoritmico.passepartout
 
-import android.Manifest
-import android.app.Notification
-import android.app.PendingIntent
 import android.content.Intent
-import android.content.pm.PackageManager
-import android.content.pm.ServiceInfo
 import android.net.VpnService
-import android.os.Build
 import android.os.IBinder
-import android.util.AtomicFile
-import androidx.core.app.NotificationChannelCompat
-import androidx.core.app.NotificationCompat
-import androidx.core.app.NotificationManagerCompat
-import androidx.core.app.ServiceCompat
-import androidx.core.content.ContextCompat
-import com.algoritmico.passepartout.business.extensions.JSON
-import com.algoritmico.passepartout.business.extensions.runCatchingNonFatal
 import com.algoritmico.passepartout.context.AppLog
 import com.algoritmico.passepartout.context.appBundle
 import com.algoritmico.passepartout.context.defaultAndroidConstants
-import com.algoritmico.passepartout.context.lastTunnelPreferences
-import com.algoritmico.passepartout.context.lastTunnelProfile
 import com.algoritmico.passepartout.context.logPreamble
-import com.algoritmico.passepartout.models.AppPreferences
-import com.algoritmico.passepartout.ui.extensions.NotificationTransferFormatter
+import com.algoritmico.passepartout.vpn.VpnServiceNotificationController
+import com.algoritmico.passepartout.vpn.VpnServiceStore
 import io.partout.NativeTunnelControllerJNI
 import io.partout.PartoutVpnServiceRuntime
 import io.partout.abi.PartoutException
-import io.partout.models.TaggedProfile
 import io.partout.models.TunnelSnapshot
-import io.partout.models.TunnelStatus
 import kotlinx.coroutines.CompletableDeferred
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.withContext
-import java.io.File
 
 class PassepartoutVpnService: VpnService() {
     private val androidConstants = defaultAndroidConstants
@@ -46,13 +27,22 @@ class PassepartoutVpnService: VpnService() {
     private val jniLogTag = androidConstants.tags.partoutJni
     private val logsSnapshots = androidConstants.tunnel.logsSnapshots
 
-    @Volatile
-    private var currentProfileName: String? = null
+    private val store by lazy {
+        VpnServiceStore(
+            logTag = logTag,
+            context = applicationContext,
+            storage = androidConstants.storage
+        )
+    }
 
-    @Volatile
-    private var shouldKeepStoppedNotification = false
-
-    private val notificationTransfer = NotificationTransferFormatter()
+    private val notifications by lazy {
+        VpnServiceNotificationController(
+            logTag = logTag,
+            service = this,
+            serviceClass = PassepartoutVpnService::class.java,
+            logsSnapshots = logsSnapshots
+        )
+    }
 
     private val runtime by lazy {
         PartoutVpnServiceRuntime(
@@ -66,10 +56,6 @@ class PassepartoutVpnService: VpnService() {
 
     private val engine = object : PartoutVpnServiceRuntime.Engine {
         private val library = PassepartoutWrapper()
-        private val lastProfileFile: File
-            get() = applicationContext.lastTunnelProfile(androidConstants.storage)
-        private val lastPreferencesFile: File
-            get() = applicationContext.lastTunnelPreferences(androidConstants.storage)
 
         override suspend fun start(
             intent: Intent?,
@@ -84,10 +70,12 @@ class PassepartoutVpnService: VpnService() {
 
             val bundle = applicationContext.appBundle()
             AppLog.d(logTag, "Bundle: $bundle")
-            updateCurrentProfileName(profileJSON)
+            notifications.updateProfileName(profileJSON)
 
             // Try preferences from intent, otherwise load last persisted
-            val preferences = readPreferences(intent)
+            val preferences = store.readPreferences(
+                intent?.getStringExtra(EXTRA_TUNNEL_PREFERENCES)
+            )
             AppLog.i(logTag, "Preferences: $preferences")
 
             // Initialize the library with the intent preferences
@@ -123,86 +111,23 @@ class PassepartoutVpnService: VpnService() {
         }
 
         override suspend fun readLastProfile(): String {
-            return withContext(Dispatchers.IO) {
-                readLastFile(lastProfileFile)
-            }
+            return store.readLastProfile()
         }
 
         override suspend fun writeLastProfile(json: String) {
-            withContext(Dispatchers.IO) {
-                writeLastFile(lastProfileFile, json)
-            }
+            store.writeLastProfile(json)
         }
 
         override suspend fun deleteLastProfile(id: String) {
-            withContext(Dispatchers.IO) {
-                runCatchingNonFatal {
-                    val json = readLastFile(lastProfileFile)
-                    val profile = JSON.decode<TaggedProfile>(json)
-                    if (profile.id != id) { return@runCatchingNonFatal }
-                    AppLog.i(logTag, "Forget last profile $id")
-                    lastProfileFile.delete()
-                }.onFailure {
-                    AppLog.w(logTag, "Unable to forget last profile", it)
-                }
-            }
+            store.deleteLastProfile(id)
         }
 
         override fun onSnapshot(snapshot: TunnelSnapshot) {
-            updateNotification(snapshot)
+            notifications.update(snapshot)
         }
 
         override fun onServiceStopped(wasRevoked: Boolean) {
-            if (wasRevoked) {
-                dismissNotification()
-                return
-            }
-            postStoppedNotification()
-        }
-
-        private fun readPreferences(intent: Intent?): AppPreferences? {
-            val intentPreferencesJSON = intent?.getStringExtra(EXTRA_TUNNEL_PREFERENCES)
-            val preferencesJSON = if (intentPreferencesJSON.isNullOrBlank()) {
-                AppLog.i(logTag, "Load last preferences")
-                runCatchingNonFatal {
-                    readLastFile(lastPreferencesFile)
-                }.onFailure {
-                    AppLog.w(logTag, "Unable to read last tunnel preferences", it)
-                }.getOrNull()
-            } else {
-                AppLog.i(logTag, "Load and persist start preferences")
-                runCatchingNonFatal {
-                    writeLastFile(lastPreferencesFile, intentPreferencesJSON)
-                }.onFailure {
-                    AppLog.w(logTag, "Unable to write last tunnel preferences", it)
-                }
-                intentPreferencesJSON
-            }
-            return preferencesJSON?.let { json ->
-                runCatchingNonFatal {
-                    JSON.decode<AppPreferences>(json)
-                }.onFailure {
-                    AppLog.w(logTag, "Unable to decode preferences JSON", it)
-                }.getOrNull()
-            }
-        }
-
-        private fun readLastFile(file: File): String {
-            return AtomicFile(file).openRead().bufferedReader(Charsets.UTF_8).use {
-                it.readText()
-            }
-        }
-
-        private fun writeLastFile(file: File, json: String) {
-            val atomicFile = AtomicFile(file)
-            val stream = atomicFile.startWrite()
-            runCatchingNonFatal {
-                stream.write(json.toByteArray(Charsets.UTF_8))
-                atomicFile.finishWrite(stream)
-            }.onFailure {
-                atomicFile.failWrite(stream)
-                throw it
-            }
+            notifications.onServiceStopped(wasRevoked)
         }
     }
 
@@ -210,16 +135,11 @@ class PassepartoutVpnService: VpnService() {
         if (intent?.action == PartoutVpnServiceRuntime.ACTION_STOP_VPN) {
             return runtime.onStartCommand(intent, flags, startId)
         }
-        shouldKeepStoppedNotification = false
-        intent?.getStringExtra(PartoutVpnServiceRuntime.EXTRA_PROFILE_JSON)?.let {
-            notificationTransfer.reset()
-            updateCurrentProfileName(it)
-        }
+        notifications.prepareStart(
+            intent?.getStringExtra(PartoutVpnServiceRuntime.EXTRA_PROFILE_JSON)
+        )
         try {
-            if (!canPostNotifications()) {
-                AppLog.w(logTag, "Starting service in foreground with notifications disabled")
-            }
-            startForegroundGracefully(createNotification(snapshot = null))
+            notifications.startForeground()
         } catch (it: SecurityException) {
             AppLog.e(logTag, "Unable to start service in foreground", it)
             return START_NOT_STICKY
@@ -231,11 +151,8 @@ class PassepartoutVpnService: VpnService() {
     }
 
     override fun onDestroy() {
-        if (!shouldKeepStoppedNotification) {
-            dismissNotification()
-        }
+        notifications.onDestroy()
         runtime.onDestroy()
-        resetNotificationState()
         super.onDestroy()
     }
 
@@ -250,199 +167,7 @@ class PassepartoutVpnService: VpnService() {
         return runtime.onBind(intent)
     }
 
-    private fun createNotification(
-        snapshot: TunnelSnapshot?,
-        isServiceStopped: Boolean = false
-    ): Notification {
-        val channelId = VPN_CHANNEL_ID
-
-        // Create a notification channel (required on Android 8.0+)
-        val channel = NotificationChannelCompat.Builder(
-            channelId,
-            NotificationManagerCompat.IMPORTANCE_LOW // low importance to avoid sound
-        )
-            .setName(getString(R.string.app_name))
-            .setDescription(getString(R.string.android_vpn_service_channel_description))
-            .setShowBadge(false)
-            .build()
-
-        NotificationManagerCompat
-            .from(this)
-            .createNotificationChannel(channel)
-
-        val title = currentProfileName ?: getString(R.string.app_name)
-
-        val builder = NotificationCompat.Builder(this, channelId)
-            .setSmallIcon(R.drawable.ic_notification_vpn)
-            .setContentTitle(title)
-            .setSubText(notificationSubText(snapshot, isServiceStopped))
-            .setCategory(NotificationCompat.CATEGORY_SERVICE)
-            .setPriority(NotificationCompat.PRIORITY_LOW)
-            .setOnlyAlertOnce(true)
-            .setOngoing(!isServiceStopped)
-            .setAutoCancel(false)
-            .addAction(
-                R.drawable.ic_notification_vpn,
-                getString(
-                    if (isServiceStopped) {
-                        R.string.global_actions_connect
-                    } else {
-                        R.string.global_actions_disconnect
-                    }
-                ),
-                if (isServiceStopped) connectPendingIntent() else disconnectPendingIntent()
-            )
-
-        val content = snapshot?.let(notificationTransfer::activeText)
-        if (content != null) {
-            builder
-                .setContentText(content)
-                .setStyle(NotificationCompat.BigTextStyle().bigText(content))
-        }
-
-        return builder.build()
-    }
-
-    private fun notificationSubText(
-        snapshot: TunnelSnapshot?,
-        isServiceStopped: Boolean
-    ): String? {
-        if (isServiceStopped) {
-            return getString(R.string.android_vpn_service_status_stopped)
-        }
-        return snapshot?.status?.let(::tunnelStatusText)
-    }
-
-    private fun tunnelStatusText(status: TunnelStatus): String {
-        val resId = when (status) {
-            TunnelStatus.inactive -> R.string.entities_tunnel_status_inactive
-            TunnelStatus.activating -> R.string.entities_tunnel_status_activating
-            TunnelStatus.active -> R.string.entities_tunnel_status_active
-            TunnelStatus.deactivating -> R.string.entities_tunnel_status_deactivating
-        }
-        return getString(resId)
-    }
-
-    private fun updateNotification(snapshot: TunnelSnapshot) {
-        if (logsSnapshots) {
-            AppLog.d(logTag, "updateNotification()")
-        }
-        val notificationManager = NotificationManagerCompat.from(this)
-        if (!canPostNotifications(notificationManager)) {
-            if (logsSnapshots) {
-                AppLog.w(logTag, "Skip VPN notification update, notifications are disabled")
-            }
-            return
-        }
-        val notification = createNotification(snapshot)
-        try {
-            notificationManager.notify(VPN_NOTIFICATION_ID, notification)
-        } catch (it: SecurityException) {
-            AppLog.w(logTag, "Unable to update VPN notification", it)
-        }
-    }
-
-    private fun postStoppedNotification() {
-        shouldKeepStoppedNotification = true
-        notificationTransfer.reset()
-        ServiceCompat.stopForeground(this, ServiceCompat.STOP_FOREGROUND_DETACH)
-
-        val notificationManager = NotificationManagerCompat.from(this)
-        if (!canPostNotifications(notificationManager)) {
-            AppLog.w(logTag, "Skip stopped VPN notification, notifications are disabled")
-            return
-        }
-        val notification = createNotification(
-            snapshot = null,
-            isServiceStopped = true
-        )
-        try {
-            notificationManager.notify(VPN_NOTIFICATION_ID, notification)
-        } catch (it: SecurityException) {
-            AppLog.w(logTag, "Unable to show stopped VPN notification", it)
-        }
-    }
-
-    private fun connectPendingIntent(): PendingIntent {
-        val intent = Intent(this, PassepartoutVpnService::class.java)
-        return PendingIntent.getService(
-            this,
-            VPN_CONNECT_REQUEST_CODE,
-            intent,
-            PendingIntent.FLAG_UPDATE_CURRENT or PendingIntent.FLAG_IMMUTABLE
-        )
-    }
-
-    private fun disconnectPendingIntent(): PendingIntent {
-        val intent = Intent(this, PassepartoutVpnService::class.java).apply {
-            action = PartoutVpnServiceRuntime.ACTION_STOP_VPN
-        }
-        return PendingIntent.getService(
-            this,
-            VPN_DISCONNECT_REQUEST_CODE,
-            intent,
-            PendingIntent.FLAG_UPDATE_CURRENT or PendingIntent.FLAG_IMMUTABLE
-        )
-    }
-
-    private fun dismissNotification() {
-        ServiceCompat.stopForeground(this, ServiceCompat.STOP_FOREGROUND_REMOVE)
-        NotificationManagerCompat
-            .from(this)
-            .cancel(VPN_NOTIFICATION_ID)
-    }
-
-    private fun startForegroundGracefully(notification: Notification) {
-        ServiceCompat.startForeground(
-            this,
-            VPN_NOTIFICATION_ID,
-            notification,
-            vpnForegroundServiceType
-        )
-    }
-
-    private val vpnForegroundServiceType: Int
-        get() {
-            return if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.UPSIDE_DOWN_CAKE) {
-                ServiceInfo.FOREGROUND_SERVICE_TYPE_SYSTEM_EXEMPTED
-            } else {
-                0
-            }
-        }
-
-    private fun canPostNotifications(
-        notificationManager: NotificationManagerCompat = NotificationManagerCompat.from(this)
-    ): Boolean {
-        if (!notificationManager.areNotificationsEnabled()) {
-            return false
-        }
-        return Build.VERSION.SDK_INT < Build.VERSION_CODES.TIRAMISU ||
-                ContextCompat.checkSelfPermission(
-                    this,
-                    Manifest.permission.POST_NOTIFICATIONS
-                ) == PackageManager.PERMISSION_GRANTED
-    }
-
-    private fun resetNotificationState() {
-        currentProfileName = null
-        notificationTransfer.reset()
-    }
-
-    private fun updateCurrentProfileName(profileJSON: String) {
-        runCatchingNonFatal {
-            JSON.decode<TaggedProfile>(profileJSON).name
-        }.onSuccess {
-            currentProfileName = it
-        }.onFailure {
-            AppLog.w(logTag, "Unable to decode VPN profile name", it)
-        }
-    }
-
     companion object {
         const val EXTRA_TUNNEL_PREFERENCES = "com.algoritmico.passepartout.extra.TUNNEL_PREFERENCES"
-        const val VPN_CHANNEL_ID = "vpn_service_channel_1"
-        const val VPN_NOTIFICATION_ID = 1
-        private const val VPN_CONNECT_REQUEST_CODE = 1000
-        private const val VPN_DISCONNECT_REQUEST_CODE = 1001
     }
 }
